fix(icons):
- Uses the correct arrow shape (not hte cursor arrow shape
- UPdates the reference images for the new icon
